start using jwt-extended

+++ resources/items.py
@@ -1,5 +1,6 @@
+from email import message
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from modules.items import ItemModule
 from typing import *
 
@@ -18,7 +19,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModule.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -54,6 +55,9 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['admin']:
+            return {'massage': "Admin privilege required"}, 401
         item = ItemModule.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -63,10 +67,13 @@ class Item(Resource):
 
 
 class Items(Resource):
-    @jwt_required()
+    @jwt_required(optional=True)
     def get(self):
-        row = ItemModule.select_all()
-        if row:
-            return {'items': [i.json() for i in row]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModule.select_all()]
+        if user_id:
+            return {'items': items}
         else:
-            return {"items": []}
+            return {'items': [item['name'] for item in items],
+                    'message': 'More details after your login'
+                    }
